pyproject.toml
[tool.poetry]
name = "real-time-web-chat"
version = "0.1.0"
description = ""
authors = ["Kylemerian <s02190176@gse.cs.msu.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.13.3"
fastapi = "^0.115.3"
sqlalchemy = "^2.0.36"
websocket = "^0.2.1"
jinja2 = "^3.1.4"
redis = "^5.1.1"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
psycopg = "^3.2.3"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
